backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.2
python-docx==1.0.1
python-jose==3.3.0
pydantic>=2.0.0
python-dotenv==1.0.0
openai==1.3.3
requests 